refactor(cli): share mentionsIn via lib + give dm/msg/ask a usage line
Move mentionsIn out of the console catalog into lib/mentions.ts so the
palette and the headless `cotal msg` import one shared helper, instead of
send.ts reaching into console internals. Set Command.usage on dm/msg/ask
so `cotal <cmd> --help` and arg errors show a clean usage line.
