+++ implementations/cli/src/index.ts
@@ -66,21 +66,24 @@ const baseCommands: Command[] = [
     kind: "command",
     name: "dm",
     group: "Mesh",
-    summary: 'send one direct message to a peer by name — dm <agent> "<text>" [--space <s>]',
+    summary: "direct-message a peer by name",
+    usage: 'dm <agent> "<text>"  [--space <s>] [--server <url>] [--creds <path>]',
     run: dm,
   },
   {
     kind: "command",
     name: "msg",
     group: "Mesh",
-    summary: 'broadcast one message to a channel — msg <channel> "<text>" [--space <s>]',
+    summary: "broadcast a message to a channel",
+    usage: 'msg <channel> "<text>"  [--space <s>] [--server <url>] [--creds <path>]',
     run: msg,
   },
   {
     kind: "command",
     name: "ask",
     group: "Mesh",
-    summary: 'anycast one message to a role (one instance answers) — ask <role> "<text>" [--space <s>]',
+    summary: "anycast a message to a role (one instance answers)",
+    usage: 'ask <role> "<text>"  [--space <s>] [--server <url>] [--creds <path>]',
     run: ask,
   },
   {
